# swagger_server/models/program.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Program(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: int=None, name: str=None, description: str=None, organization_id: int=None, duration: int=None):  # noqa: E501
        """Program - a model defined in Swagger

        :param id: The id of this Program.  # noqa: E501
        :type id: int
        :param name: The name of this Program.  # noqa: E501
        :type name: str
        :param description: The description of this Program.  # noqa: E501
        :type description: str
        :param organization_id: The organization_id of this Program.  # noqa: E501
        :type organization_id: int
        :param duration: The duration of this Program.  # noqa: E501
        :type duration: int
        """
        self.swagger_types = {
            'id': int,
            'name': str,
            'description': str,
            'organization_id': int,
            'duration': int
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name',
            'description': 'description',
            'organization_id': 'organization_id',
            'duration': 'duration'
        }
        self._id = id
        self._name = name
        self._description = description
        self._organization_id = organization_id
        self._duration = duration

    @classmethod
    def from_dict(cls, dikt) -> 'Program':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Program of this Program.  # noqa: E501
        :rtype: Program
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this Program.


        :return: The id of this Program.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this Program.


        :param id: The id of this Program.
        :type id: int
        """

        self._id = id

    @property
    def name(self) -> str:
        """Gets the name of this Program.


        :return: The name of this Program.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this Program.


        :param name: The name of this Program.
        :type name: str
        """

        self._name = name

    @property
    def description(self) -> str:
        """Gets the description of this Program.


        :return: The description of this Program.
        :rtype: str
        """
        return self._description

    @description.setter
    def description(self, description: str):
        """Sets the description of this Program.


        :param description: The description of this Program.
        :type description: str
        """

        self._description = description

    @property
    def organization_id(self) -> int:
        """Gets the organization_id of this Program.


        :return: The organization_id of this Program.
        :rtype: int
        """
        return self._organization_id

    @organization_id.setter
    def organization_id(self, organization_id: int):
        """Sets the organization_id of this Program.


        :param organization_id: The organization_id of this Program.
        :type organization_id: int
        """

        self._organization_id = organization_id

    @property
    def duration(self) -> int:
        """Gets the duration of this Program.


        :return: The duration of this Program.
        :rtype: int
        """
        return self._duration

    @duration.setter
    def duration(self, duration: int):
        """Sets the duration of this Program.


        :param duration: The duration of this Program.
        :type duration: int
        """

        self._duration = duration